refactor : 비디오, 오디오 컴포넌트로 분리
