+++ components/Coretime/WebRTCMedia/index.ts
@@ -1,0 +1,2 @@
+export { default as WebRTCAudio } from './WebRTCAudio';
+export { default as WebRTCVideo } from './WebRTCVideo';
